backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
open-clip-torch==2.24.0
torch>=2.9.0
torchvision>=0.20.0
Pillow>=11.0.0
qdrant-client==1.9.1
imagehash==4.3.1
httpx==0.27.0
python-dotenv==1.0.1
cloudinary==1.40.0
apscheduler==3.10.4
pydantic>=2.9.0
pydantic-settings>=2.5.0
